+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fixed Validate Example files in:
   - ConfigureAutoMountPoints-FromCalculator.ps1
   - ConfigureAutoMountPoints-Manual.ps1
+  - ConfigureDatabases-FromCalculator.ps1
   - InstallExchange.ps1
   - WaitForADPrep.ps1